+++ app/client/src/apis/write.ts
@@ -1,4 +1,4 @@
-import { Activity } from "../types/activity";
+import { Activity } from "../types/write";
 import { API } from "../types/api";
 
 export async function submitActivity(data: Activity) {
